Rename client_is_human to is-client-human
